- change aspect from a scaled int to a double so we can handle the wider    range of aspect ratios we get from ffmpeg files.
 - add container_aspect to title struct (always zero except for DVDs
   when it's the aspect from the VTSI). To handle broken French DVDs,
   make HB complain & use the container aspect if it's different from
   the aspect computed from the video PAR.

 - fix ScanFunc's job template init so that it doesn't think the only
   legal aspect ratios are 16:9 & 4:3.

 - hb_reduce wouldn't reduce any fraction where both terms were equal and
   prime (e.g., 2/2, 3/3, 5/5, etc. would not become 1/1). Recoded it using
   Euclid's Algorithm so it always works and is faster.


git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@1616 b64f7644-9d1e-0410-96f1-a4d463321fa5

+++ libhb/hb.h
@@ -90,7 +90,7 @@ int hb_detect_comb( hb_buffer_t * buf, int width, int height, int color_equal, i
 
 void          hb_get_preview( hb_handle_t *, hb_title_t *, int,
                               uint8_t * );
-void          hb_set_size( hb_job_t *, int ratio, int pixels );
+void          hb_set_size( hb_job_t *, double ratio, int pixels );
 void          hb_set_anamorphic_size( hb_job_t *,
                 int *output_width, int *output_height,
                 int *output_par_width, int *output_par_height);
